Move the existing Markdown code into the package
