+++ src/pincushion/flask/filters.py
@@ -4,7 +4,9 @@ import functools
 import re
 
 import markdown
+from markdown.extensions import Extension
 from markdown.extensions.smarty import SmartyExtension
+from markdown.preprocessors import Preprocessor
 
 
 def title_markdown(md):
@@ -19,6 +21,37 @@ def title_markdown(md):
 
     res = markdown.markdown(md, extensions=[SmartyExtension()])
     return res.replace('<p>', '').replace('</p>', '')
+
+
+class AutoLinkPreprocessor(Preprocessor):
+    """
+    Preprocessor that converts anything that looks like a URL into a link.
+    """
+    def run(self, lines):
+        new_lines = []
+        for line in lines:
+            for u in re.findall(r'(https?://[^\s]+?)(?:\s|$)', line):
+                line = line.replace(u, f'<{u}>')
+            new_lines.append(line)
+        return new_lines
+
+
+class AutoLinkExtension(Extension):
+    def extendMarkdown(self, md, md_globals):
+        md.registerExtension(self)
+        md.preprocessors.add(
+            'inline_urls',
+            AutoLinkPreprocessor(md),
+            '>normalize_whitespace'
+        )
+
+
+def description_markdown(md):
+    """Renders a Markdown string as HTML for use in a bookmark description."""
+    return markdown.markdown(md, extensions=[
+        SmartyExtension(),
+        AutoLinkExtension()
+    ])
 
 
 def cmp(x, y):
